+++ R/build-readme.R
@@ -26,7 +26,12 @@ build_rmd <- function(files, path = ".", output_options = list(), ..., quiet = T
   # Ensure rendering github_document() doesn't generate HTML file
   output_options$html_preview <- FALSE
 
-  paths <- path(pkg$path, files)
+  paths <- files
+
+  abs_files <- is_absolute_path(files)
+
+  paths[!abs_files] <- path(pkg$path, files[!abs_files])
+
   for (path in paths) {
     cli::cli_alert_info("Building {.path {path}}")
     callr::r_safe(
@@ -45,16 +50,15 @@ build_rmd <- function(files, path = ".", output_options = list(), ..., quiet = T
 #' @export
 build_readme <- function(path = ".", quiet = TRUE, ...) {
   pkg <- as.package(path)
-  inst <- path(pkg$path, "inst")
-  dirs <- c(pkg$path, if (dir_exists(inst)) inst)
 
-  readme_path <- grep(
-    ignore.case = TRUE, value = TRUE,
-    "readme[.]rmd",
-    path_file(dir_ls(dirs))
-  )
+  readme_path <- path_abs(dir_ls(ignore.case = TRUE, regexp = "(inst/)?readme[.]rmd", recurse = 1, type = "file"))
+
   if (length(readme_path) == 0) {
-    return(invisible())
+    rlang::abort("Can't find a 'README.Rmd' or 'inst/README.Rmd' file.")
+  }
+
+  if (length(readme_path) > 1) {
+    rlang::abort("Can't have both a 'README.Rmd' and 'inst/README.Rmd' file.")
   }
 
   build_rmd(readme_path, path = path, quiet = quiet, ...)